requirements.txt
numpy==2.4.2
pandas==3.0.1
PyYAML==6.0.3
